energy bar
added energy bar
replaced dummy element parents with recipes
fixed double combining. Combine was called twice because the other
element fired the event too. Now we check if the mouse is down for the
element before we set the mousedown.
Energy is lost by combining and half of energy by splitting elements